Added InteractiveBrokers provider. (click for detail)
Added support for interactive brokers. To make use of it you need the TWS installed, compile the InteractiveBrokers solution under the Providers folder, and set the ProviderAssembly property in the GUI app.config to be IBProviderService.

The IBProviderService support Level1 and TimeAndSales (but not Level2 yet) or T&S + Level1 combined. Use the settings in the Symbol dictionary to determine you data feed level. Testing InteractiveBrokers Provider plugin.