Added TimeZone and DisplayTimeZone to the symbol dictionary. (click for details)
The DisplayTimeZone can have the values of UTC, Local, or Exchange.
If you set it to Exchange then thes symbol will use the TimeZone
property to determine which time zone to use for display.
The TimeZone property should be set to the time zone of the exchange
where the symbol is traded. The time zone names are the same
as the Windows time zones without abbreviations like "Eastern Standard Time".
Also, for testing purposed TickZoom supports time zone names in the
form of UTC-2 or UTC+3 for specific UTC offsets. However, these never
account for daylight savings times. So you normally always want to use
the standard time zone names.