Added strategy flag IsActiveOrdersModified and send of entry vs. exit orders from kernel. (click for details)
This means that the orders will only be resent to the broker if any of the orders has been modified. Also, the kernel automatically sends either a list of entry orders only if the current position is flat or else only the exit orders if there is a long/short position. It also automatically adjust the exit order sizes to the current position any time the position size changes.